refactor: swap form (#248)
