+++ apps/app.mento.org/app/components/swap/swap-form/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./swap-form";
