+++ apps/app.mento.org/app/components/swap/swap-form/swap-submit-button.tsx
@@ -1,0 +1,148 @@
+"use client";
+
+import { IconLoading } from "@repo/ui";
+import { Button } from "@repo/ui";
+import { ConnectButton } from "@repo/web3";
+
+import type { TokenWithBalance } from "@repo/web3";
+
+interface SwapSubmitButtonProps {
+  isConnected: boolean;
+  hasAmount: boolean;
+  tokenInSymbol: string;
+  tokenOutSymbol: string;
+  quote: string | undefined;
+  errors: { amount?: { message?: string } };
+  isButtonLoading: boolean;
+  isApproveTxLoading: boolean;
+  isApprovalProcessing: boolean;
+  tradingLimitError: string | null;
+  balanceError: string | null;
+  isTradingSuspended: boolean;
+  isSuspensionCheckLoading: boolean;
+  isError: boolean;
+  canQuote: boolean;
+  shouldApprove: string | boolean;
+  allTokenOptions: TokenWithBalance[];
+}
+
+export function SwapSubmitButton({
+  isConnected,
+  hasAmount,
+  tokenInSymbol,
+  tokenOutSymbol,
+  quote,
+  errors,
+  isButtonLoading,
+  isApproveTxLoading,
+  isApprovalProcessing,
+  tradingLimitError,
+  balanceError,
+  isTradingSuspended,
+  isSuspensionCheckLoading,
+  isError,
+  canQuote,
+  shouldApprove,
+  allTokenOptions,
+}: SwapSubmitButtonProps) {
+  if (!isConnected) {
+    return (
+      <ConnectButton
+        size="lg"
+        text="Connect"
+        fullWidth
+        shouldShowAddress={false}
+      />
+    );
+  }
+
+  return (
+    <Button
+      data-testid={defineButtonLocator({
+        balanceError,
+        tradingLimitError,
+        isTradingSuspended,
+        shouldApprove,
+        tokenInSymbol,
+        tokenOutSymbol,
+      })}
+      className="mt-auto w-full"
+      size="lg"
+      clipped="lg"
+      type="submit"
+      disabled={
+        !hasAmount ||
+        !tokenOutSymbol ||
+        !tokenInSymbol ||
+        !quote ||
+        !!(errors.amount && errors.amount.message !== "Amount is required") ||
+        isButtonLoading ||
+        isApproveTxLoading ||
+        isApprovalProcessing ||
+        !!tradingLimitError ||
+        !!balanceError ||
+        isTradingSuspended ||
+        isSuspensionCheckLoading ||
+        (isError && hasAmount && canQuote)
+      }
+    >
+      {isButtonLoading ? (
+        <IconLoading />
+      ) : !tokenInSymbol ? (
+        "Select token to sell"
+      ) : !tokenOutSymbol ? (
+        "Select token to buy"
+      ) : isTradingSuspended ? (
+        `Trading suspended for ${tokenInSymbol} -> ${tokenOutSymbol}`
+      ) : tradingLimitError ? (
+        "Swap exceeds trading limits"
+      ) : balanceError ? (
+        "Insufficient balance"
+      ) : isError && hasAmount && canQuote ? (
+        "Unable to fetch quote"
+      ) : errors.amount?.message &&
+        errors.amount?.message !== "Amount is required" ? (
+        errors.amount?.message
+      ) : isApproveTxLoading || isApprovalProcessing ? (
+        <IconLoading />
+      ) : shouldApprove ? (
+        `Approve ${allTokenOptions.find((t) => t.symbol === tokenInSymbol)?.symbol || tokenInSymbol}`
+      ) : (
+        "Swap"
+      )}
+    </Button>
+  );
+}
+
+function defineButtonLocator({
+  balanceError,
+  tradingLimitError,
+  isTradingSuspended,
+  shouldApprove,
+  tokenInSymbol,
+  tokenOutSymbol,
+}: {
+  balanceError: string | null;
+  tradingLimitError: string | null;
+  isTradingSuspended: boolean;
+  shouldApprove: string | boolean;
+  tokenInSymbol: string;
+  tokenOutSymbol: string;
+}) {
+  switch (true) {
+    case Boolean(isTradingSuspended):
+      return "tradingSuspendedButton";
+    case Boolean(balanceError && !tradingLimitError):
+      return "insufficientBalanceButton";
+    case Boolean(tradingLimitError):
+      return "swapsExceedsTradingLimitButton";
+    case Boolean(shouldApprove && tokenInSymbol && tokenOutSymbol):
+      return "approveButton";
+    case !tokenInSymbol:
+      return "selectTokenToSellButton";
+    case !tokenOutSymbol:
+      return "selectTokenToBuyButton";
+    default:
+      return "swapButton";
+  }
+}
