+++ apps/app.mento.org/app/components/swap/swap-form/use-swap-form.tsx
@@ -1,0 +1,725 @@
+"use client";
+
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useForm, useWatch } from "react-hook-form";
+import { toast } from "sonner";
+
+import type { TokenSymbol } from "@mento-protocol/mento-sdk";
+import {
+  CELO_EXPLORER,
+  chainIdToChain,
+  confirmViewAtom,
+  formatBalance,
+  formatWithMaxDecimals,
+  formValuesAtom,
+  fromWeiRounded,
+  getTokenDecimals,
+  logger,
+  MIN_ROUNDED_VALUE,
+  parseAmount,
+  parseAmountWithDefault,
+  type SwapFormValues,
+  toWei,
+  useAccountBalances,
+  useApproveTransaction,
+  useOptimizedSwapQuote,
+  useSwapAllowance,
+  useTokenOptions,
+  useTradablePairs,
+  useTradingLimits,
+  useTradingSuspensionCheck,
+} from "@repo/web3";
+import { useAccount, useChainId } from "@repo/web3/wagmi";
+import { useAtom } from "jotai";
+import { OctagonAlert } from "lucide-react";
+
+import { defaultEmptyBalances, formSchema, type FormValues } from "./types";
+
+export function useSwapForm() {
+  const { address, isConnected } = useAccount();
+  const chainId = useChainId() ?? 42220;
+  const [formValues, setFormValues] = useAtom(formValuesAtom);
+  const [, setConfirmView] = useAtom(confirmViewAtom);
+  const [isApprovalProcessing, setIsApprovalProcessing] = useState(false);
+
+  const { data: balancesFromHook } = useAccountBalances({ address, chainId });
+  const balances = balancesFromHook || defaultEmptyBalances;
+
+  const { allTokenOptions } = useTokenOptions(undefined, balancesFromHook);
+
+  const amountRef = useRef<HTMLInputElement>(null);
+  const quoteRef = useRef<HTMLInputElement>(null);
+  const prevTradingSuspensionErrorRef = useRef<string | null>(null);
+  const suspensionToastIdRef = useRef<string | number | null>(null);
+
+  const form = useForm<FormValues>({
+    resolver: zodResolver(formSchema),
+    defaultValues: {
+      amount: formValues?.amount || "",
+      quote: formValues?.quote || "",
+      tokenInSymbol: formValues?.tokenInSymbol || "CELO",
+      tokenOutSymbol: formValues?.tokenOutSymbol || "USDm",
+      slippage: formValues?.slippage || "0.5",
+    },
+    mode: "onChange",
+  });
+
+  const tokenInSymbol = useWatch({
+    control: form.control,
+    name: "tokenInSymbol",
+  }) as TokenSymbol;
+  const tokenOutSymbol = useWatch({
+    control: form.control,
+    name: "tokenOutSymbol",
+  }) as TokenSymbol;
+  const amount = useWatch({ control: form.control, name: "amount" });
+  const formQuote = useWatch({ control: form.control, name: "quote" });
+
+  // Token balances
+  const fromTokenBalance = useMemo(() => {
+    const balanceValue = balances[tokenInSymbol as keyof typeof balances];
+    const balance = formatBalance(
+      balanceValue,
+      getTokenDecimals(tokenInSymbol, chainId),
+    );
+    return formatWithMaxDecimals(balance || "0.00");
+  }, [balances, tokenInSymbol, chainId]);
+
+  const toTokenBalance = useMemo(() => {
+    const balanceValue = balances[tokenOutSymbol as keyof typeof balances];
+    const balance = fromWeiRounded(
+      balanceValue,
+      getTokenDecimals(tokenOutSymbol, chainId),
+    );
+    return formatWithMaxDecimals(balance || "0.00");
+  }, [balances, tokenOutSymbol, chainId]);
+
+  // Trading limits
+  const { data: limits, isLoading: limitsLoading } = useTradingLimits(
+    tokenInSymbol,
+    tokenOutSymbol,
+    chainId,
+  );
+
+  // Trading suspension
+  const {
+    isSuspended: isTradingSuspended,
+    isLoading: isSuspensionCheckLoading,
+  } = useTradingSuspensionCheck(tokenInSymbol, tokenOutSymbol);
+
+  // ── Validation ──────────────────────────────────────────────────────
+
+  const validateBalance = useCallback(
+    (value: string) => {
+      if (!value || !tokenInSymbol) return true;
+      if (value === "0." || value === "0") return true;
+
+      const parsedAmount = parseAmount(value);
+      if (!parsedAmount) return true;
+
+      if (parsedAmount.lte(MIN_ROUNDED_VALUE) && !parsedAmount.isZero()) {
+        return "Amount too small";
+      }
+
+      const tokenInfo = allTokenOptions.find((t) => t.symbol === tokenInSymbol);
+      if (!tokenInfo) return "Invalid token";
+
+      const balance = balances[tokenInSymbol as keyof typeof balances];
+      if (typeof balance === "undefined") return "Balance unavailable";
+
+      const amountInWei = toWei(parsedAmount, tokenInfo.decimals || 18);
+      const balanceInWei = parseAmountWithDefault(balance, "0");
+
+      if (
+        amountInWei.gt(0) &&
+        (balanceInWei.isZero() || balanceInWei.lt(amountInWei))
+      ) {
+        return "Insufficient balance";
+      }
+
+      return true;
+    },
+    [balances, tokenInSymbol, allTokenOptions],
+  );
+
+  const checkTradingLimitViolation = useCallback(
+    (
+      numericAmount: number,
+      numericQuote: number,
+      limits: NonNullable<ReturnType<typeof useTradingLimits>["data"]>,
+      tokenInSymbol: string,
+      tokenOutSymbol: string,
+    ) => {
+      const { L0, L1, LG, tokenToCheck } = limits;
+
+      let amountToCheck: number;
+      let exceeds = false;
+      let limit = 0;
+      let total = 0;
+      let timestamp = 0;
+      let exceededTier: "L0" | "L1" | "LG" | null = null;
+      let isImplicitLimit = false;
+
+      if (tokenToCheck === tokenInSymbol) {
+        amountToCheck = numericAmount;
+        if (LG?.maxIn && amountToCheck > LG.maxIn) {
+          exceeds = true;
+          limit = LG.maxIn;
+          timestamp = LG.until || 0;
+          exceededTier = "LG";
+          total = LG.total || 0;
+        } else if (L1?.maxIn && amountToCheck > L1.maxIn) {
+          exceeds = true;
+          limit = L1.maxIn;
+          timestamp = L1.until || 0;
+          exceededTier = "L1";
+          total = L1.total || 0;
+        } else if (L0?.maxIn && amountToCheck > L0.maxIn) {
+          exceeds = true;
+          limit = L0.maxIn;
+          timestamp = L0.until || 0;
+          exceededTier = "L0";
+          total = L0.total || 0;
+        }
+      } else if (tokenToCheck === tokenOutSymbol) {
+        amountToCheck = numericQuote;
+
+        if (LG?.maxOut && amountToCheck > LG.maxOut) {
+          exceeds = true;
+          limit = LG.maxOut;
+          timestamp = LG.until || 0;
+          exceededTier = "LG";
+          total = LG.total || 0;
+        } else if (L1?.maxOut && amountToCheck > L1.maxOut) {
+          exceeds = true;
+          limit = L1.maxOut;
+          timestamp = L1.until || 0;
+          exceededTier = "L1";
+          total = L1.total || 0;
+        } else if (L0?.maxOut && amountToCheck > L0.maxOut) {
+          exceeds = true;
+          limit = L0.maxOut;
+          timestamp = L0.until || 0;
+          exceededTier = "L0";
+          total = L0.total || 0;
+        }
+        isImplicitLimit = true;
+      }
+
+      if (exceeds) {
+        if (isImplicitLimit) {
+          if (exceededTier === "LG") {
+            return `Cannot buy more than ${limit.toLocaleString()} ${tokenToCheck}. This exceeds the global trading limit.`;
+          } else {
+            const date = new Date(timestamp * 1000).toLocaleString();
+            const timeframe = exceededTier === "L0" ? "5min" : "1d";
+            return `Cannot buy more than ${limit.toLocaleString()} ${tokenToCheck} within ${timeframe}. The limit will reset to ${total.toLocaleString()} ${tokenToCheck} at ${date}.`;
+          }
+        } else {
+          if (exceededTier === "LG") {
+            return `The ${tokenToCheck} amount exceeds the global trading limit of ${limit.toLocaleString()} ${tokenToCheck}.`;
+          } else {
+            const date = new Date(timestamp * 1000).toLocaleString();
+            const timeframe = exceededTier === "L0" ? "5min" : "1d";
+            return `The ${tokenToCheck} amount exceeds the current trading limit of ${limit.toLocaleString()} ${tokenToCheck} within ${timeframe}. It will be reset again to ${total.toLocaleString()} ${tokenToCheck} at ${date}.`;
+          }
+        }
+      }
+
+      return null;
+    },
+    [],
+  );
+
+  const validateLimits = useCallback(
+    async (value: string) => {
+      if (!value || limitsLoading || !limits || !limits.tokenToCheck)
+        return true;
+      if (value === "0." || value === "0") return true;
+
+      const parsedAmount = parseAmount(value);
+      if (!parsedAmount) return true;
+
+      const numericAmount = Number(parsedAmount.toString()) || 0;
+      const numericQuote = Number(formQuote) || 0;
+
+      const violation = checkTradingLimitViolation(
+        numericAmount,
+        numericQuote,
+        limits,
+        tokenInSymbol,
+        tokenOutSymbol,
+      );
+
+      return violation || true;
+    },
+    [
+      limitsLoading,
+      limits,
+      checkTradingLimitViolation,
+      tokenInSymbol,
+      tokenOutSymbol,
+      formQuote,
+    ],
+  );
+
+  const validateAmount = useCallback(
+    async (value: string) => {
+      const balanceCheck = validateBalance(value);
+      if (balanceCheck !== true) return balanceCheck;
+
+      const limitsCheck = await validateLimits(value);
+      if (limitsCheck !== true) return limitsCheck;
+
+      return true;
+    },
+    [validateBalance, validateLimits],
+  );
+
+  // ── Handlers ────────────────────────────────────────────────────────
+
+  const handleReverseTokens = () => {
+    const currentTokenInSymbol = form.getValues("tokenInSymbol");
+    const currentTokenOutSymbol = form.getValues("tokenOutSymbol");
+    const currentAmount = form.getValues("amount");
+
+    form.setValue("tokenInSymbol", currentTokenOutSymbol);
+    form.setValue("tokenOutSymbol", currentTokenInSymbol);
+    form.setValue("amount", currentAmount);
+    form.setValue("quote", "");
+  };
+
+  const handleUseMaxBalance = () => {
+    let maxAmountInWei =
+      balances[tokenInSymbol as keyof typeof balances] || "0";
+    const decimals = getTokenDecimals(tokenInSymbol, chainId);
+
+    if (tokenInSymbol === "CELO") {
+      const gasReserveWei = BigInt("10000000000000000"); // 0.01 CELO
+      const balance = BigInt(maxAmountInWei);
+      if (balance > gasReserveWei) {
+        maxAmountInWei = (balance - gasReserveWei).toString();
+      }
+    }
+
+    const formattedAmount = formatBalance(maxAmountInWei, decimals);
+    const formattedAmountWithMaxDecimals = formatWithMaxDecimals(
+      formattedAmount,
+      4,
+      false,
+    );
+    form.setValue("amount", formattedAmountWithMaxDecimals);
+
+    if (tokenInSymbol === "CELO") {
+      toast.success("Max balance used", {
+        duration: 5000,
+        description: () => <>0.01 CELO reserved for transaction fees</>,
+        icon: <OctagonAlert strokeWidth={1.5} size={18} className="mt-0.5" />,
+      });
+    }
+  };
+
+  // ── Derived state ───────────────────────────────────────────────────
+
+  const { errors } = form.formState;
+  const hasAmount =
+    !!amount &&
+    amount !== "" &&
+    amount !== "0" &&
+    amount !== "0." &&
+    Number(amount) > 0;
+
+  const [tradingLimitError, setTradingLimitError] = useState<string | null>(
+    null,
+  );
+  const [balanceError, setBalanceError] = useState<string | null>(null);
+
+  const tradingSuspensionError = useMemo(() => {
+    if (!isTradingSuspended) return null;
+    return `Trading temporarily paused for ${tokenInSymbol} -> ${tokenOutSymbol}. Unable to determine accurate exchange rate now. Please try again later.`;
+  }, [isTradingSuspended, tokenInSymbol, tokenOutSymbol]);
+
+  const canQuote =
+    !!hasAmount && !errors.amount && !limitsLoading && !isTradingSuspended;
+
+  // ── Quote ───────────────────────────────────────────────────────────
+
+  const {
+    isFetching: quoteFetching,
+    quote,
+    rate,
+    isError,
+    fromTokenUSDValue,
+    toTokenUSDValue,
+  } = useOptimizedSwapQuote(
+    canQuote ? amount : "",
+    tokenInSymbol,
+    tokenOutSymbol,
+  );
+
+  // ── Side effects ────────────────────────────────────────────────────
+
+  useEffect(() => {
+    setTradingLimitError(null);
+  }, [amount, quote]);
+
+  useEffect(() => {
+    const checkBalance = async () => {
+      if (!hasAmount || !tokenInSymbol) {
+        setBalanceError(null);
+        return;
+      }
+      const balanceCheck = validateBalance(amount);
+      if (balanceCheck !== true) {
+        setBalanceError(balanceCheck);
+      } else {
+        setBalanceError(null);
+      }
+    };
+    checkBalance();
+  }, [amount, hasAmount, tokenInSymbol, validateBalance]);
+
+  useEffect(() => {
+    if (!hasAmount || !limits || limitsLoading) return;
+
+    const numericAmount = Number(parseAmount(amount) ?? 0);
+    const numericQuote = Number(quote || 0);
+
+    const violation = checkTradingLimitViolation(
+      numericAmount,
+      numericQuote,
+      limits,
+      tokenInSymbol,
+      tokenOutSymbol,
+    );
+
+    setTradingLimitError((v) => (v === violation ? v : violation));
+  }, [
+    amount,
+    quote,
+    limits,
+    limitsLoading,
+    tokenInSymbol,
+    tokenOutSymbol,
+    hasAmount,
+    checkTradingLimitViolation,
+  ]);
+
+  useEffect(() => {
+    if (tradingLimitError) {
+      toast.error(tradingLimitError, { duration: 20000 });
+    }
+  }, [tradingLimitError]);
+
+  useEffect(() => {
+    const prevError = prevTradingSuspensionErrorRef.current;
+    const hasError = tradingSuspensionError !== null;
+    const errorChanged = prevError !== tradingSuspensionError;
+
+    if (hasError && errorChanged) {
+      if (suspensionToastIdRef.current) {
+        toast.dismiss(suspensionToastIdRef.current);
+      }
+      suspensionToastIdRef.current = toast.error(tradingSuspensionError, {
+        duration: 20000,
+      });
+    } else if (prevError !== null && !hasError) {
+      if (suspensionToastIdRef.current) {
+        toast.dismiss(suspensionToastIdRef.current);
+        suspensionToastIdRef.current = null;
+      }
+    }
+
+    prevTradingSuspensionErrorRef.current = tradingSuspensionError;
+  }, [tradingSuspensionError]);
+
+  const isLoading =
+    quoteFetching && canQuote && !tradingLimitError && !limitsLoading;
+
+  const prevTokenPairRef = useRef<{
+    tokenInSymbol: TokenSymbol | undefined;
+    tokenOutSymbol: TokenSymbol | undefined;
+  }>({ tokenInSymbol: undefined, tokenOutSymbol: undefined });
+
+  const [isWaitingForQuote, setIsWaitingForQuote] = useState(false);
+
+  useEffect(() => {
+    const tokensChanged =
+      prevTokenPairRef.current.tokenInSymbol !== tokenInSymbol ||
+      prevTokenPairRef.current.tokenOutSymbol !== tokenOutSymbol;
+
+    if (tokensChanged) {
+      prevTokenPairRef.current = { tokenInSymbol, tokenOutSymbol };
+      if (hasAmount && !!tokenInSymbol && !!tokenOutSymbol) {
+        setIsWaitingForQuote(true);
+      }
+    }
+
+    if (isTradingSuspended && isWaitingForQuote) {
+      setIsWaitingForQuote(false);
+    }
+
+    if (
+      isWaitingForQuote &&
+      quote &&
+      quote !== "0" &&
+      Number(quote) > 0 &&
+      !quoteFetching
+    ) {
+      setIsWaitingForQuote(false);
+    }
+  }, [
+    tokenInSymbol,
+    tokenOutSymbol,
+    hasAmount,
+    isWaitingForQuote,
+    quote,
+    quoteFetching,
+    isTradingSuspended,
+  ]);
+
+  const isButtonLoading = useMemo(
+    () =>
+      !isTradingSuspended &&
+      (quoteFetching || isWaitingForQuote) &&
+      hasAmount &&
+      !!tokenInSymbol &&
+      !!tokenOutSymbol,
+    [
+      quoteFetching,
+      isWaitingForQuote,
+      hasAmount,
+      tokenInSymbol,
+      tokenOutSymbol,
+      isTradingSuspended,
+    ],
+  );
+
+  const sellUSDValue = useMemo(() => {
+    return tokenInSymbol === "USDm" ? amount || "0" : fromTokenUSDValue || "0";
+  }, [tokenInSymbol, amount, fromTokenUSDValue]);
+
+  const buyUSDValue = useMemo(() => {
+    return tokenOutSymbol === "USDm"
+      ? formQuote || "0"
+      : toTokenUSDValue || "0";
+  }, [tokenOutSymbol, formQuote, toTokenUSDValue]);
+
+  const amountInWei = useMemo(() => {
+    if (!tokenInSymbol) return "0";
+    return amount
+      ? toWei(amount, getTokenDecimals(tokenInSymbol, chainId)).toFixed(0)
+      : "0";
+  }, [amount, tokenInSymbol, chainId]);
+
+  // ── Allowance & approval ────────────────────────────────────────────
+
+  const { skipApprove } = useSwapAllowance({
+    chainId,
+    tokenInSymbol,
+    tokenOutSymbol,
+    approveAmount: amountInWei,
+    address,
+  });
+
+  const { sendApproveTx, isApproveTxLoading } = useApproveTransaction({
+    chainId,
+    tokenInSymbol,
+    tokenOutSymbol,
+    amountInWei,
+    accountAddress: address,
+    onSuccess: (receipt) => {
+      logger.info("Approval transaction confirmed");
+      const chain = chainIdToChain[chainId];
+      const explorerUrl = chain?.blockExplorers?.default.url;
+      const explorerName =
+        chain?.blockExplorers?.default?.name || CELO_EXPLORER.name;
+      toast.success(
+        <>
+          <h4>Approve Successful</h4>
+          <span className="mt-2 block text-muted-foreground">
+            Token allowance for swap approved
+          </span>
+          {explorerUrl && (
+            <a
+              href={`${explorerUrl}/tx/${receipt.transactionHash}`}
+              target="_blank"
+              rel="noopener noreferrer"
+              className="text-muted-foreground underline"
+            >
+              View Transaction on {explorerName}
+            </a>
+          )}
+        </>,
+      );
+      setIsApprovalProcessing(false);
+
+      const currentFormValues = form.getValues();
+      const formData: SwapFormValues = {
+        ...currentFormValues,
+        slippage: formValues?.slippage || currentFormValues.slippage || "0.5",
+        isAutoSlippage: formValues?.isAutoSlippage,
+        deadlineMinutes: formValues?.deadlineMinutes,
+        tokenInSymbol,
+        tokenOutSymbol,
+        buyUSDValue,
+        sellUSDValue,
+      };
+
+      setFormValues(formData);
+      setConfirmView(true);
+    },
+  });
+
+  useEffect(() => {
+    if (quote !== undefined) {
+      const formattedQuote = formatWithMaxDecimals(quote, 4, false);
+      if (formQuote !== formattedQuote) {
+        form.setValue("quote", formattedQuote, { shouldValidate: true });
+      }
+    }
+  }, [quote, form, formQuote]);
+
+  useEffect(() => {
+    if (
+      errors.amount?.message &&
+      errors.amount.message !== "Invalid input" &&
+      errors.amount.message !== "Amount is required" &&
+      hasAmount
+    ) {
+      toast.error(errors.amount.message);
+    }
+  }, [errors.amount, hasAmount, amount]);
+
+  // ── Submit ──────────────────────────────────────────────────────────
+
+  const onSubmit = async (values: FormValues) => {
+    try {
+      if (!skipApprove && sendApproveTx) {
+        setIsApprovalProcessing(true);
+        logger.info("Approval needed, sending approve transaction");
+        const hash = await sendApproveTx();
+
+        if (!hash) {
+          setIsApprovalProcessing(false);
+        }
+
+        logger.info("Waiting for approval transaction", {
+          hash,
+        });
+      } else {
+        const formData: SwapFormValues = {
+          ...values,
+          slippage: formValues?.slippage || form.getValues("slippage") || "0.5",
+          isAutoSlippage: formValues?.isAutoSlippage,
+          deadlineMinutes: formValues?.deadlineMinutes,
+          tokenInSymbol,
+          tokenOutSymbol,
+          buyUSDValue,
+          sellUSDValue,
+        };
+        setFormValues(formData);
+        setConfirmView(true);
+      }
+    } catch (error) {
+      logger.error("Error in swap form submission", error);
+      setIsApprovalProcessing(false);
+    }
+  };
+
+  const shouldApprove =
+    !skipApprove && hasAmount && quote && !isLoading && !balanceError;
+
+  // ── Token pair validation ───────────────────────────────────────────
+
+  const { data: fromTokenTradablePairs } = useTradablePairs(tokenInSymbol);
+  const { data: toTokenTradablePairs } = useTradablePairs(tokenOutSymbol);
+
+  const [lastChangedToken, setLastChangedToken] = useState<
+    "from" | "to" | null
+  >(null);
+
+  useEffect(() => {
+    if (!tokenInSymbol || !tokenOutSymbol || !lastChangedToken) return;
+
+    const isValidPair =
+      fromTokenTradablePairs?.includes(tokenOutSymbol) ||
+      toTokenTradablePairs?.includes(tokenInSymbol);
+
+    if (!isValidPair) {
+      if (lastChangedToken === "from") {
+        form.setValue("tokenOutSymbol", "", { shouldValidate: false });
+      } else if (lastChangedToken === "to") {
+        form.setValue("tokenInSymbol", "", { shouldValidate: false });
+      }
+      setLastChangedToken(null);
+    }
+  }, [
+    tokenInSymbol,
+    tokenOutSymbol,
+    fromTokenTradablePairs,
+    toTokenTradablePairs,
+    lastChangedToken,
+    form,
+  ]);
+
+  // ── Return ──────────────────────────────────────────────────────────
+
+  return {
+    form,
+    isConnected,
+
+    // Refs
+    amountRef,
+    quoteRef,
+
+    // Token symbols & options
+    tokenInSymbol,
+    tokenOutSymbol,
+    allTokenOptions,
+
+    // Watched values
+    amount,
+    formQuote,
+
+    // Balances
+    fromTokenBalance,
+    toTokenBalance,
+
+    // USD values
+    sellUSDValue,
+    buyUSDValue,
+
+    // Quote data
+    quote,
+    rate,
+    isError,
+    canQuote,
+
+    // Validation
+    validateAmount,
+    errors,
+    hasAmount,
+    tradingLimitError,
+    balanceError,
+    isTradingSuspended,
+    isSuspensionCheckLoading,
+
+    // Loading
+    isLoading,
+    isButtonLoading,
+    isApproveTxLoading,
+    isApprovalProcessing,
+
+    // Approval
+    shouldApprove,
+    skipApprove,
+
+    // Handlers
+    handleReverseTokens,
+    handleUseMaxBalance,
+    onSubmit,
+    setLastChangedToken,
+  };
+}
